src/requirements.txt
tensorflow
keras
numpy
pillow
